stage1/requirements.txt
charset-normalizer==2.1.1
colorama==0.4.6
filelock==3.8.0
fire==0.4.0
ftfy==6.1.1
huggingface-hub==0.11.0
idna==3.4
numpy
open-clip-torch==2.7.0
packaging==21.3
patool==1.12
Pillow==9.3.0
Pygments==2.13.0
pyparsing==3.0.9
python-fire==0.1.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
termcolor==2.1.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
wincertstore==0.2
